test(processing): genuinely exercise streaming-partial fallback suppression
The prior `streaming_partial_origin_local_no_shift` passed
`site_position = None`, so `site_position.is_none()` short-circuited the
stage-2 gate open and the test never actually covered the "first-pass
detection SUCCEEDS and thus SUPPRESSES the fallback" path its name implied.

Rework it as `streaming_partial_first_pass_success_suppresses_fallback`:
build a genuinely PARTIAL index that holds the origin-local wall's full
placement + representation chain (so the first-pass detect succeeds) plus the
IfcSite ENTITY but NOT its forward-referenced placement, and pass
`site_position = Some`. Both fallback arms are then closed by the success
itself. The IfcSite placement stays at the large national-grid coordinate, so
`scan_placement_bounds` would return a large offset if the placement-bounds
fallback were reached — the test asserts the resolved offset is instead the
partial pass's (0,0,0), and directly asserts the first pass returns Some.

Split the now-four-case test module into a sibling `stream_meta_tests.rs`
(exempt from the module-size ratchet, per its `*_tests.rs` rule) so the
production module stays ~200 lines. Tests drive only the public surface.
